% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwelnet-package.R
\docType{package}
\name{cooper-package}
\alias{cooper-package}
\alias{_PACKAGE}
\alias{fwelnet-package}
\title{cooper: Cooperative Penalized Regression}
\description{
A method for fitting cause-specific penalized Cox regression models based on the feature-weighted elastic net (fwelnet) introduced by Tay, J. K., Aghaeepour, N., Hastie, T., and Tibshirani, R., (2020) \href{https://arxiv.org/abs/2006.01395}{arXiv:2006.01395}.
}
\author{
\strong{Maintainer}: Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

Authors:
\itemize{
  \item Kenneth Tay \email{kjytay@stanford.edu}
  \item Robert Tibshirani
}

}
\keyword{internal}
