% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwelnet_mt.R
\name{extract_beta_history}
\alias{extract_beta_history}
\title{Extract beta coefficients across cooper iterations}
\usage{
extract_beta_history(object)
}
\arguments{
\item{object}{Object of class \code{"cooper"}, see \code{\link[=cooper]{cooper()}}.}
}
\value{
A \link{data.table}.
}
\description{
Extract beta coefficients across cooper iterations
}
