% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwelnet-package.R
\docType{package}
\name{fwelnet-package}
\alias{fwelnet-package}
\alias{_PACKAGE}
\title{fwelnet: Feature-Weighted Elastic Net}
\description{
A method for fitting the entire regularization path of the feature-weighted elastic net (fwelnet) for linear and logistic regression. See Tay, J. K., Aghaeepour, N., Hastie, T., and Tibshirani, R., (2020) \href{https://arxiv.org/abs/2006.01395}{arXiv:2006.01395} for details.
}
\keyword{internal}
