% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwelnet_mt.R
\name{fwelnet_mt_cox}
\alias{fwelnet_mt_cox}
\title{fwelnet for multi-task learning}
\usage{
fwelnet_mt_cox(
  data,
  causes = 1:2,
  mt_max_iter = 5,
  alpha = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame or matrix holding predictors and outcome,
with outcome variables assumed to be named `"time"` and `""status`.}

\item{causes}{Integer vector indicating causes, e.g. `1:2` for two causes.}

\item{mt_max_iter}{`[5]` number of mt-iterations to perform. Will break early
if no change in per-cause beta vector between iterations is detected.}

\item{alpha}{Passed to [`glmnet()`] and [`fwelnet()`]}

\item{verbose}{Display informative message on the state of the mt fit.}

\item{...}{Passed to [`fwelnet()`[]}
}
\value{
A `list` containing per=cause beta matrices for each iteration step
}
\description{
Survival-focused implementation of the fwelnet-multitask algorithm described
in Algorithm 2 on page 13 (Tay et. al. 2020)
}
