% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nll_cox}
\alias{nll_cox}
\title{Cox NLL shortcut}
\usage{
nll_cox(x, y, beta)
}
\arguments{
\item{x}{n x p matrix}

\item{y}{Vector of length n}

\item{beta}{p x nlam matrix}
}
\description{
Get pred/nll from survival::coxph(), 0 iterations, coefs stay the same.
Vectorized over nlam cols in beta
}
