% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cooper.R
\name{predictRisk.cooper}
\alias{predictRisk.cooper}
\title{predictRisk for Cooper models}
\usage{
predictRisk.cooper(object, newdata, times, cause, use_initial_fit = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{cooper}, containing two cause-specific Cox model fits.}

\item{newdata}{New data to predict on.}

\item{times}{Event times to evaluate at}

\item{cause}{Event of interest (\code{1} or \code{2}).}

\item{use_initial_fit}{(\code{FALSE}) If \code{TRUE}, uses the initial \code{glmnet} fit of the cooper algorithm.}

\item{...}{Unused}
}
\value{
Matrix of dimension \code{nrow(newdata)} x \code{length(times)}
}
\description{
predictRisk for Cooper models
}
\examples{
if (requireNamespace("riskRegression")) {
library(riskRegression)

data(pbc, package = "survival")
pbc <- na.omit(pbc)
xtrain <- pbc[1:270, -1]
xtest <- pbc[270:276, -1]

set.seed(12)
cooperfit <- cooper(
  xtrain, mt_max_iter = 3,
  stratify_by_status = TRUE,
  standardize = FALSE # Currently required to ensure correct results
)

predictRisk(cooperfit, xtest, cause = 1, times = c(1180.5, 1788))

}
}
