% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-cr.R
\name{sim_cr_coxexp}
\alias{sim_cr_coxexp}
\title{Simple competing risk simulation}
\usage{
sim_cr_coxexp(
  n = 100,
  p = 4,
  beta1 = runif(p, -1, 1),
  beta2 = beta1,
  lambda1 = 0.1,
  lambda2 = lambda1,
  lambda_c = 0.1,
  r = 0
)
}
\arguments{
\item{n, p}{Number of observations and number of features.}

\item{beta1, beta2}{Coefficient vectors of the Cox-exponential model.}

\item{lambda1, lambda2}{Baseline hazard constants of the models for event 1 and 2.}

\item{lambda_c}{Baseline hazard constant for the uninformative censoring distribution.}

\item{r}{Correlation between features}
}
\description{
Simple competing risk simulation
}
\examples{
sim_cr_coxexp()
}
