% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwelnet_mt.R
\name{fwelnet_mt_cox}
\alias{fwelnet_mt_cox}
\title{Fit fwelnet for Multi-Task Learning}
\usage{
fwelnet_mt_cox(
  data,
  causes = 1:2,
  mt_max_iter = 5,
  z_method = "original",
  stratify_by_status = FALSE,
  alpha = 1,
  verbose = FALSE,
  t = 1,
  a = 0.5,
  thresh = 0.001,
  nfolds = 10,
  include_mt_beta_history = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame or matrix holding predictors and outcome,
with outcome variables assumed to be named \code{"time"} and \code{"status"}.}

\item{causes}{(Unused for now) Integer vector indicating causes, e.g. \code{1:2} for two causes.}

\item{mt_max_iter}{\verb{[5]} number of mt-iterations to perform. Will break early
if no change in per-cause beta vector between iterations is detected.
If set to \code{0}, no \code{fwelnet} iteration will be performed and the returned
coefficients will be the result of fitting a cause-specific \code{glmnet}.}

\item{z_method}{\code{("original")} Either assign \code{z1} to be informed by \code{beta2}
of the current iteration (default behavior, as described in Algorithm 2
in Tay et. al. 2020), or \code{"aligned"} to have both \code{z1} and \code{z2} be informed
by \code{beta2} and \code{beta1} from the previous iteration step respectively.}

\item{stratify_by_status}{(\code{FALSE}) If \code{TRUE}, the internal cross-validation folds are sampled by the \code{status}
variable, leading to folds with roughly equal proportions of events as in \code{data}.
This is helpful in cases where many censored observations in folds lead to errors during model fits.}

\item{alpha}{\code{(1)} Passed to \code{\link[=glmnet]{glmnet()}} and \code{\link[=fwelnet]{fwelnet()}}.}

\item{verbose}{Display informative message on the state of the mt fit.}

\item{t}{The initial step size for \code{theta} backtracking line search
(default value is 1).}

\item{a}{The factor by which step size is decreased in \code{theta}
backtracking line search (default value is 0.5).}

\item{thresh}{If the mean/median objective function does not decrease by at
least this factor, we terminate the optimization early. Default is 1e-4.}

\item{nfolds}{(\code{10}) Passed to \code{\link{cv.fwelnet}} and \code{\link{cv.glmnet}} respectively.}

\item{include_mt_beta_history}{\verb{[FALSE]} If \code{TRUE}, output \code{list} includes
components \code{beta1} and \code{beta2}, matrices of dimensions \code{p} x \code{mt_iter_max + 1}
containing coefficient vectors for causes 1 and 2 for each multi-task iteration,
with the first column corresponding to the original \code{cv.glmnet} solution.}

\item{...}{Passed to \code{\link[=fwelnet]{fwelnet()}}}
}
\value{
A \code{list} containing per-cause beta matrices for each iteration step
}
\description{
Survival-focused implementation of the fwelnet-multitask algorithm described
in Algorithm 2 on page 13 (Tay et. al. 2020)
}
